Add message changeset to emque-producing
This change adds a message changeset and changes payload generator.
Clients can use this functionality by including the message
functionality with Emque::Producing.message(:with_changeset => true).
This will provide the basic message functionality along with the ability
to pass in an original and updated object that can be diffed such that
the message includes details about the delta of the objects.

The design for the inclusion affords the ability for clients to define
this behavior on a message by message basis.