Add config option to turn publishing on or off
This option is being added to enable clients to turn the actual
publishing of the message to the broker on or off. The obvious example
is in the case of CI environments that may not support or be running the
broker, we need not fail on unpublished messages so long as they are
still being validated.