Add ability for clients to override changeset keys
This commit is being submitted to enable clients to address issues of
needing to normalize message changeset contents across disparate message
producers. Specifically, when Producer A has a slightly different
naming convention than that of Producer B, this commit allows either
Producer A or Producer B to translate a given set of attributes in the
object changeset to be consistent.

The specs are specifically closely aligned with how the gem is
used in production by passing in a hash of attrs with strings as keys.
The specs will fail with symbol keys because the code only
translates keys that are an exact match. This PR also sanitizes the
changeset attrs and ensures that both the keys and values are strings
for the translation.

The addition of the deep copy is required in order to maintain thread
safety by serializing the original and updated changesets and returning
new translated copies as opposed to mutating the existing hashes.