Make optional publishing part of the message, not RabbitMQ publishing
This change comes full circle, leaving the publisher as it was. This
means that the publisher will always know whether the message was sent
or not. But, the message can decide whether that's worth raising an
error for.

The tradeoff is perhaps a performance cost. Non-mandatory messages
that don't use the confirms would be a lot faster for non-critical
messages.